# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Новостного портала",
        "description": "Участники: Веденин Иван, Тишкина Алина, Федор Пережогин, Александр Сергеев",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://news-portal.local/api",
            "description": "API"
        }
    ],
    "paths": {
...
        "/save": {
            "get": {
                "tags": [
                    "Save"
                ],
                "summary": "Get list of saves",
                "description": "Returns list of saves",
                "operationId": "getSaveList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Save"
                ],
                "summary": "Store new save",
                "description": "Returns save data",
                "operationId": "storeSave",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/save/{id}": {
            "get": {
                "tags": [
                    "Save"
                ],
                "summary": "Get save information",
                "description": "Returns save data",
                "operationId": "getSaveById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Save id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Save"
                ],
                "summary": "Update existing save",
                "description": "Returns updated save data",
                "operationId": "updateSave",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Save id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Save"
                ],
                "summary": "Delete existing save",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteSave",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Save id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/subscribes": {
            "get": {
                "tags": [
                    "Subscribes"
                ],
                "summary": "Get list of subscribes",
                "description": "Returns list of subscribes",
                "operationId": "getSubscribesList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscribes"
                ],
                "summary": "Store new subscribes",
                "description": "Returns subscribes data",
                "operationId": "storeSubscribes",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/subscribes/{id}": {
            "get": {
                "tags": [
                    "Subscribes"
                ],
                "summary": "Get subscribes information",
                "description": "Returns subscribes data",
                "operationId": "getSubscribesById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subscribes id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subscribes"
                ],
                "summary": "Update existing subscribes",
                "description": "Returns updated subscribes data",
                "operationId": "updateSubscribes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subscribes id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subscribes"
                ],
                "summary": "Delete existing subscribes",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteSubscribes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subscribes id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}